requirements.txt
fastapi
uvicorn[standard]
transformers
sentencepiece
Pillow
torch
protobuf
python-multipart